requirements.txt
annotated-types==0.7.0
anyio==4.9.0
blinker==1.9.0
certifi==2025.4.26
click==8.2.1
Deprecated==1.2.18
distro==1.9.0
Flask==3.1.1
Flask-Limiter==3.12
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.10.0
limits==5.3.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
openai==1.86.0
ordered-set==4.1.0
packaging==25.0
pillow==11.2.1
pydantic==2.11.5
pydantic_core==2.33.2
Pygments==2.19.1
python-dotenv==1.1.0
rich==13.9.4
shortuuid==1.0.13
sniffio==1.3.1
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.0
Werkzeug==3.1.3
wrapt==1.17.2
